requirements.txt
asgiref==3.2.7
Django==2.1.1
django-taggit==1.3.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1